+++ matzipback/src/main/java/com/matzip/matzipback/matzipList/command/domain/aggregate/MyListMatzip.java
@@ -9,9 +9,9 @@ import lombok.NoArgsConstructor;
 public class MyListMatzip {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int listMatzipSeq;
-    private int listSeq;
-    private int restaurantSeq;
+    private long listMatzipSeq;
+    private long listSeq;
+    private long restaurantSeq;
     private String listMatzipComment;
 
 }
